fs: Extract some syscalls
