+++ src/compat/posix/fs/dvfs/stat.c
@@ -1,0 +1,32 @@
+/**
+ * @file
+ *
+ * @data 12 oct. 2015
+ * @author: Anton Bondarev
+ */
+
+#include <fcntl.h>
+
+#include <fs/dvfs.h>
+
+int stat(const char *path, struct stat *buf) {
+	struct lookup l;
+	int res = dvfs_lookup(path, &l);
+	struct dentry *d = l.item;
+
+	if (res < 0)
+		return res;
+
+	assert(d);
+
+	*buf = (struct stat) {
+		.st_dev     = (d && d->d_sb && d->d_sb->bdev) ? d->d_sb->bdev->id : 0,
+		.st_ino     = d->d_inode->i_no,
+		.st_nlink   = 1,
+		.st_size    = (d && d->d_inode) ? d->d_inode->length : 0,
+		.st_blksize = 512,
+		.st_mode    = d->flags & S_IFMT,
+	};
+
+	return 0;
+}
